% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DendrometerSensors.R
\docType{data}
\name{DendrometerSensors}
\alias{DendrometerSensors}
\title{DendrometerSensors}
\format{A data.frame with columns
\describe{
\item{column}{Each column contains measurements from different dendrometer sensor.}
\item{row}{Time of recording (in hours)}

}}
\source{
Agricultural plant-based IoT company,
}
\usage{
DendrometerSensors
}
\description{
Dendrometers are measurement devices that continuously monitor size variation in plant organs like stems, branches and fruits. 
Data from dendrometers can be used for monitoring plants’ daily water status and accordingly make automated irrigation decisions.
The dataset includes data of 100 sensors from 120 hourly readings (5 days).
}
\keyword{datasets}
