% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{EigenCentrality}
\alias{EigenCentrality}
\title{EigenCentrality}
\usage{
EigenCentrality(mat)
}
\arguments{
\item{mat}{dataset in the format of data.table, data.frame or matrix which contains continuous variables from the same scale.}
}
\value{
A vector of numeric scores per each variable.
}
\description{
Method for calculating multivariate anomaly scores,
based on Eigenvector centrality, which is a popular network centrality measure.
The anomaly scores are calculated based on distance matrix estimate.
}
\examples{
data <- DendrometerSensors

#Create score per sensor
EigenCentrality.score<-EigenCentrality(data)

#View scores with MultiNav function
anomaly.scores<- cbind(id=as.numeric(colnames(data)), EigenCentrality.score)
MultiNav(data,type = "scores", scores = anomaly.scores)

}
