% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{extract_sd0}
\alias{extract_sd0}
\title{extract_sd0}
\usage{
extract_sd0(data)
}
\arguments{
\item{data}{Matrix, data.table or data.frame. Containing continuous variables.}
}
\value{
Returns a data.table, data.frame or matrix (in the same format as the input data).
}
\description{
Data pre-processing utility function. Removes from a given dataset all rows and columns that have 0 standard deviation.
}
\details{
Function that can be used as pre-processing step.
}
\examples{
data <- DendrometerSensors

# dimentions before pre-processing
dim(data)
new_data<-extract_sd0(data)
# dimentions after pre-processing
dim(new_data)

}
\seealso{
\code{\link[MultiNav]{MultiNav}}, \code{\link[MultiNav]{Calc_uni_matrix}}, \code{\link[MultiNav]{Calc_quantiles_matrix}},  \code{\link[MultiNav]{sliding_window_scores}}
}
