% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LambsWeight.R
\docType{data}
\name{LambsWeight}
\alias{LambsWeight}
\title{LambsWeight}
\format{A data.frame with columns
\describe{
\item{column}{Each column contains weight measurement of a different lamb.} 
\item{row}{Time of recording (in hours)}
}}
\source{
Volcani center, ARO PLF lab. Data was collected from a novel system which contains electronic scales and drinking behavior sensor, designed for automatic small-ruminant monitoring.
}
\usage{
LambsWeight
}
\description{
Daily weight estimates of individual lambs in a small herd.
The dataset contains data of 80 daily weight estimates for 39 lambs.
}
\keyword{datasets}
