% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{sliding_window_scores}
\alias{sliding_window_scores}
\title{sliding_window_scores}
\usage{
sliding_window_scores(data, window, scores = NULL)
}
\arguments{
\item{data}{Matrix, data.table or data.frame. Containing continuous variables.}

\item{window}{Integer, indicating the size of the temporal window for which scores will be calculated (in a sliding window manner).}

\item{scores}{Character, specifies which scores set to calculate. If not specified, default univariate scores (quantiles 0.25, 0.5, 0.75 and MAD) will be calculated. Additionally two default multivariate score sets are supported, named: "mv" and "T2_Variations".}
}
\value{
Returns a data.table, data.frame or matrix (in the same format as the input data).
}
\description{
Data pre-processing utility function. Calculates sliding window scores for default scores sets.
}
\details{
Function that can be used as pre-processing step, to calculate sliding window scores per sensor (variable).
}
\examples{
data <- LambsWeight

# Calculate sliding window univariate scores.
scores<-sliding_window_scores(data,14)
head(scores)

# Calculate sliding window scores with "T2_Variations" multivariate scores set.
scores<-sliding_window_scores(data,14,scores="T2_Variations")
head(scores)

# Calculate sliding window scores with "mv" multivariate scores set.
scores<-sliding_window_scores(data,14,scores="mv")
head(scores)

}
\seealso{
\code{\link[MultiNav]{MultiNav}}, \code{\link[MultiNav]{Calc_uni_matrix}}, \code{\link[MultiNav]{Calc_quantiles_matrix}}, \code{\link[MultiNav]{extract_sd0}}
}
