% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{T2}
\alias{T2}
\title{T2}
\usage{
T2(mat, cov = "pearson")
}
\arguments{
\item{mat}{Matrix, data.table or data.frame. Containing continuous variables.}

\item{cov}{Character, specifies the method of cov estimate "pearson", "spearman", "mcd75", "mcd50", "mve","Srivastava-Du","croux_ollerer"}
}
\value{
A vector of numeric scores per observation (row).
}
\description{
Hotelling T2 scoring method for calculating multivariate anomaly scores.
}
\examples{
data <- LambsWeight[58:62,]

#Create score per sensor
T2.score<-T2(t(data))

#View scores with MultiNav function
anomaly.scores<- cbind(id=as.numeric(colnames(data)), T2.score)
MultiNav(data,type = "scores", scores = anomaly.scores)

}
\references{
ROUSSEEUW, P. J., et al. Robustbase: basic robust statistics. R package version 0.4-5, R-project. org/package= robustbase, 2009.

HOTELLING, HAROLD. Multivariate quality control. Techniques of statistical analysis. McGraw-Hill, New York, 1947.
}
\seealso{
\code{\link[MultiNav]{MultiNav}}, \code{\link[MultiNav]{T2_mcd50}}, \code{\link[MultiNav]{T2_mcd75}}, \code{\link[MultiNav]{T2_mve}}, \code{\link[MultiNav]{T2_croux_ollerer}}, \code{\link[MultiNav]{T2_srivastava_du}}
}
