% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{T2_croux_ollerer}
\alias{T2_croux_ollerer}
\title{croux_ollerer}
\usage{
T2_croux_ollerer(mat)
}
\arguments{
\item{mat}{Matrix, data.table or data.frame. Containing continuous variables.}
}
\value{
A vector of numeric scores per observation (row).
}
\description{
Method for calculating multivariate anomaly scores.
A variation of popular Hotelling T2,
calculated based on covariance estimate with Croux-Ollerer method.
}
\examples{
data <- LambsWeight[58:62,]

#Create score per sensor
T2_croux_ollerer.score<-T2_croux_ollerer(t(data))

#View scores with MultiNav function
anomaly.scores<- cbind(id=as.numeric(colnames(data)), T2_croux_ollerer.score)
MultiNav(data,type = "scores", scores = anomaly.scores)

}
\references{
HOTELLING, HAROLD. Multivariate quality control. Techniques of statistical analysis. McGraw-Hill, New York, 1947.

CROUX, Christophe; ÖLLERER, Viktoria. Robust and sparse estimation of the inverse covariance matrix using rank correlation measures. In: Recent Advances in Robust Statistics: Theory and Applications. Springer, New Delhi, 2016. p. 35-55.
}
\seealso{
\code{\link[MultiNav]{MultiNav}},\code{\link[MultiNav]{T2}}, \code{\link[MultiNav]{T2_mcd50}}, \code{\link[MultiNav]{T2_mcd75}}, \code{\link[MultiNav]{T2_mve}},  \code{\link[MultiNav]{T2_srivastava_du}}
}
