% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{T2_CrouxOllerer}
\alias{T2_CrouxOllerer}
\title{Croux-Ollerer}
\usage{
T2_CrouxOllerer(mat)
}
\arguments{
\item{mat}{dataset in the format of data.table, data.frame or matrix which contains continuous variables from the same scale.}
}
\value{
A vector of numeric scores per observation (row).
}
\description{
Method for calculating multivariate anomaly scores.
A variation of popular Hotelling T2,
calculated based on covariance estimate with Croux-Ollerer method.
}
\examples{
data <- LambsWeight[58:62,]

#Create score per sensor
T2_CrouxOllerer.score<-T2_CrouxOllerer(t(data))

#View scores with MultiNav function
anomaly.scores<- cbind(id=as.numeric(colnames(data)), T2_CrouxOllerer.score)
MultiNav(data,type = "scores", scores = anomaly.scores)

}
\references{
Montgomery, Douglas C. Introduction to statistical quality control. John Wiley & Sons, 2007.

CROUX, Christophe; ÖLLERER, Viktoria. Robust and sparse estimation of the inverse covariance matrix using rank correlation measures. In: Recent Advances in Robust Statistics: Theory and Applications. Springer, New Delhi, 2016. p. 35-55.
}
