% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{T2_mcd50}
\alias{T2_mcd50}
\title{T2_mcd50}
\usage{
T2_mcd50(mat)
}
\arguments{
\item{mat}{Matrix, data.table or data.frame. Containing continuous variables.}
}
\value{
A vector of numeric scores per observation (row).
}
\description{
Method for calculating multivariate anomaly scores.
A variation of popular Hotelling T2,
calculated based on covariance estimate with mcd method where
the determinant is minimized based on 50 percent subset of the data.
}
\examples{
data <- LambsWeight[58:62,]
# Create score per sensor
T2_mcd50.score<-T2_mcd50(t(data))

# View scores with MultiNav function
anomaly.scores<- cbind(id=as.numeric(colnames(data)), T2_mcd50.score)
MultiNav(data,type = "scores", scores = anomaly.scores)

}
\references{
ROUSSEEUW, P. J., et al. Robustbase: basic robust statistics. R package version 0.4-5, URL http://CRAN. R-project. org/package= robustbase, 2009.

HOTELLING, HAROLD. Multivariate quality control. Techniques of statistical analysis. McGraw-Hill, New York, 1947.

Wilcox, Rand R. Introduction to robust estimation and hypothesis testing. Academic Press, 2012.
}
\seealso{
\code{\link[MultiNav]{MultiNav}}, \code{\link[MultiNav]{T2}}, \code{\link[MultiNav]{T2_mcd75}}, \code{\link[MultiNav]{T2_mve}}, \code{\link[MultiNav]{T2_srivastava_du}}, \code{\link[MultiNav]{T2_croux_ollerer}}
}
