% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{T2_mcd75}
\alias{T2_mcd75}
\title{T2_mcd75}
\usage{
T2_mcd75(mat)
}
\arguments{
\item{mat}{dataset of continuous variables from the same scale in matrix format.}
}
\value{
A vector of numeric scores per observation (row).
}
\description{
Method for calculating multivariate anomaly scores.
A variation of popular Hotelling T2,
calculated based on covariance estimate with mcd method where
the determinant is minimized based on 75 percent subset of the data.
}
\examples{
data <- LambsWeight[58:62,]

#Create score per sensor
T2_mcd75.score<-T2_mcd75(t(data))

#View scores with MultiNav function
anomaly.scores<- cbind(id=as.numeric(colnames(data)), T2_mcd75.score)
MultiNav(data,type = "scores", scores = anomaly.scores)

}
\references{
ROUSSEEUW, P. J., et al. Robustbase: basic robust statistics. R package version 0.4-5, URL http://CRAN. R-project. org/package= robustbase, 2009.

Montgomery, Douglas C. Introduction to statistical quality control. John Wiley & Sons, 2007.

Wilcox, Rand R. Introduction to robust estimation and hypothesis testing. Academic Press, 2012.
}
