% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{T2_mve}
\alias{T2_mve}
\title{T2_mve}
\usage{
T2_mve(mat)
}
\arguments{
\item{mat}{Matrix, data.table or data.frame. Containing continuous variables.}
}
\value{
A vector of numeric scores per observation (row).
}
\description{
Method for calculating multivariate anomaly scores.
A variation of popular Hotelling T2,
calculated based on covariance estimate with mve method.
}
\examples{
data <- LambsWeight[58:62,]

#Create score per sensor
T2_mve.score<-T2_mve(t(data))

#View scores with MultiNav function
anomaly.scores<- cbind(id=as.numeric(colnames(data)), T2_mve.score)
MultiNav(data,type = "scores", scores = anomaly.scores)

}
\references{
HOTELLING, HAROLD. Multivariate quality control. Techniques of statistical analysis. McGraw-Hill, New York, 1947.

Wilcox, Rand R. Introduction to robust estimation and hypothesis testing. Academic Press, 2012.
}
\seealso{
\code{\link[MultiNav]{MultiNav}}, \code{\link[MultiNav]{T2}}, \code{\link[MultiNav]{T2_mcd50}}, \code{\link[MultiNav]{T2_mcd75}}, \code{\link[MultiNav]{T2_srivastava_du}}, \code{\link[MultiNav]{T2_croux_ollerer}}
}
