% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{T2_srivastava_du}
\alias{T2_srivastava_du}
\title{T2_srivastava_du}
\usage{
T2_srivastava_du(mat)
}
\arguments{
\item{mat}{Matrix, data.table or data.frame. Containing continuous variables.}
}
\value{
A vector of numeric scores per observation (row).
}
\description{
Method for calculating multivariate anomaly scores.
A variation of popular Hotelling T2,
calculated based on covariance estimate with Srivastava-Du method.
}
\examples{
data <- LambsWeight[58:62,]

#Create score per sensor
T2_srivastava_du.score<-T2_srivastava_du(t(data))

#View scores with MultiNav function
anomaly.scores<- cbind(id=as.numeric(colnames(data)), T2_srivastava_du.score)
MultiNav(data,type = "scores", scores = anomaly.scores)

}
\references{
HOTELLING, HAROLD. Multivariate quality control. Techniques of statistical analysis. McGraw-Hill, New York, 1947.

Srivastava, Muni S., and Meng Du. "A test for the mean vector with fewer observations than the dimension." Journal of Multivariate Analysis 99.3 (2008): 386-402
}
\seealso{
\code{\link[MultiNav]{MultiNav}}, \code{\link[MultiNav]{T2}}, \code{\link[MultiNav]{T2_mcd50}}, \code{\link[MultiNav]{T2_mcd75}}, \code{\link[MultiNav]{T2_mve}}, \code{\link[MultiNav]{T2_croux_ollerer}}
}
