% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{T2_SrivastavaDu}
\alias{T2_SrivastavaDu}
\title{T2_SrivastavaDu}
\usage{
T2_SrivastavaDu(mat)
}
\arguments{
\item{mat}{dataset in the format of data.table, data.frame or matrix which contains continuous variables from the same scale.}
}
\value{
A vector of numeric scores per observation (row).
}
\description{
Method for calculating multivariate anomaly scores.
A variation of popular Hotelling T2,
calculated based on covariance estimate with Srivastava-Du method.
}
\examples{
data <- LambsWeight[58:62,]

#Create score per sensor
T2_SrivastavaDu.score<-T2_SrivastavaDu(t(data))

#View scores with MultiNav function
anomaly.scores<- cbind(id=as.numeric(colnames(data)), T2_SrivastavaDu.score)
MultiNav(data,type = "scores", scores = anomaly.scores)

}
\references{
Montgomery, Douglas C. Introduction to statistical quality control. John Wiley & Sons, 2007.

Srivastava, Muni S., and Meng Du. "A test for the mean vector with fewer observations than the dimension." Journal of Multivariate Analysis 99.3 (2008): 386-402
}
