% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{Calc_quantiles_matrix}
\alias{Calc_quantiles_matrix}
\title{Calc_quantiles_matrix}
\usage{
Calc_quantiles_matrix(data, quantiles = c(0.05, 0.25, 0.75, 0.95))
}
\arguments{
\item{data}{Matrix, data.table or data.frame. Containing continuous variables. A unique numeric row name should be assigned to each observation and will be used as unique identifier (id). If row names were not assigned, row number will be used as id. A unique numeric column name should be assigned to each variable.}

\item{quantiles}{Optional argument. Gives option to specify custom quantiles instead of the default: quantiles = c(0.05, 0.25, 0.75, 0.95).}
}
\value{
Returns a data.frame
}
\description{
Creates collection of five selected quantiles performed on each row (observation) of the data.
The default quantiles are: 5, 25, 50, 75 and 95.
Minimum and maximum of each row are also included in the output dataset.
The ‘quantiles matrix’ is created in the format needed as input for MultiNav's functional boxplot chart.
}
\examples{
library(MultiNav)
data <- DendrometerSensors
quantiles_matrix<-Calc_quantiles_matrix(data)
head(quantiles_matrix)

#Custom quantiles example:
quantiles_matrix<-Calc_quantiles_matrix(data, quantiles = c(0.02,0.2,0.8,0.98))
head(quantiles_matrix)

}
\seealso{
\code{\link[MultiNav]{MultiNav}}, \code{\link[MultiNav]{Calc_uni_matrix}}, \code{\link[MultiNav]{extract_sd0}}, \code{\link[MultiNav]{sliding_window_scores}}
}
