% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{Calc_uni_matrix}
\alias{Calc_uni_matrix}
\title{Calc_uni_matrix}
\usage{
Calc_uni_matrix(data, op = FALSE)
}
\arguments{
\item{data}{dataset in the format of data.table, data.frame or matrix which contains continuous variables from the same scale. A unique numeric row name should be assigned to each observation and will be used as unique identifier (id). A unique numeric column name should be assigned to each variable.}

\item{op}{Logical, if TRUE, boxcox transformation is calculated for each descriptive statistic. In addition outlier flag is added per each descriptive statistic, with value 1 to indicate outlier, 0 to indicate inliner. The outlier flag is set according to 3 sigma control limits.}
}
\value{
Returns a data.frame
}
\description{
Creates collection of univariate descriptive statistics performed on the values
of each variable (sensor): minimum, quantile 25, median, quantile 75, maximum, mean, standard deviation,
mad, skewness and kurtosis.
}
\examples{
library(MultiNav)
data <- DendrometerSensors
uni_matrix<-Calc_uni_matrix(data)
head(uni_matrix)

uni_matrix<-Calc_uni_matrix(data, op=TRUE)
head(uni_matrix)

}
