% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{extract_sd0}
\alias{extract_sd0}
\title{extract_sd0}
\usage{
extract_sd0(data)
}
\arguments{
\item{data}{dataset in the format of data.table, data.frame or matrix which contains continuous variables.}
}
\value{
Returns a data.table, data.frame or matrix (in th same format as the input data).
}
\description{
Data pre-processing utility function. Removes from a given dataset all rows and columns that have 0 standard deviation.
}
\details{
Function that can be used as pre-processing step.
}
\examples{
data <- DendrometerSensors
extract_sd0(data)

}
