% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiNav.R
\name{MultiNav}
\alias{MultiNav}
\title{MultiNav}
\usage{
MultiNav (data = NULL, type = NULL, scores = NULL, quantiles = NULL, window=NULL, ...)
}
\arguments{
\item{data}{A continuous dataset to be analyzed, where rows correspond to equally spaced time observations and columns correspond to different variables (such as readings from different sensors). The data input can be object of class matrix, data.table or data.frame.}

\item{type}{Character, indicating which display method to invoke. Currently, one interactive screen is supported, named \code{"scores"} and several base charts are supported, named: \code{"line"}, \code{"scatter"}, \code{"network"}, \code{"functional_box"}.}

\item{scores}{Dataset, containing calculated anomaly scores per sensor (dataset variables). If not specified, default univariate scores (quantiles 0.25, 0.5, 0.75 and MAD) will be calculated. Additionally two default multivariate score sets are supported, named: "mv" and "T2_Variations".}

\item{quantiles}{Dataset, containing custom quantiles used for creating the functional box plot background.}

\item{window}{Integer, indicating the size of the temporal window for which scores will be calculated. If omitted, the default is one temporal window (that contains all the data).}

\item{, , , }{Other arguments passed on to methods.}
}
\description{
MultiNav() function for calling MultiNav's interactive visualization screens and base charts.
}
\examples{
library(MultiNav)

# "scores" display with default univariate scores set
data <- DendrometerSensors
MultiNav(data, type = "scores")

# "scores" display with "mv" default multivariate scores set
data<-LambsWeight
window<-3
data.3D<-data[(dim(data)[2]-window+1):dim(data)[2],]
MultiNav(data.3D,type = "scores", scores = "mv")

# "scores" display with "T2_Variations" default multivariate scores set
MultiNav(data.3D,type = "scores", scores = "T2_Variations")

#Passing custom scores
quantile_25 <- apply(data,2,quantile,probs=0.25,na.rm = TRUE)
median <- apply(data,2,median,na.rm = TRUE)
quantile_75 <- apply(data,2,quantile,probs=0.75,na.rm = TRUE)
mad<- apply(data,2,mad,na.rm = TRUE)
anomaly.scores<- cbind(id=as.numeric(colnames(data)),
                       quantile_25,median,quantile_75,mad)
head(anomaly.scores)
MultiNav(data,type = "scores", scores = anomaly.scores)

#Passing custom quantiles
quantiles_matrix<-Calc_quantiles_matrix(data)
head(quantiles_matrix)
MultiNav(data,type = "scores", quantiles=quantiles_matrix)

# Examine historical scores per sensor (click on sensor score to view historical scores).
MultiNav(data,type = "scores", scores = "mv",window=3)


}
