% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{T2}
\alias{T2}
\title{T2}
\usage{
T2(mat, cov = "pearson")
}
\arguments{
\item{mat}{dataset of continuous variables from the same scale in matrix format.}

\item{cov}{Character, specifies the method of cov estimate "pearson", "spearman", "mcd75", "mcd50", "mve","Srivastava-Du","Croux-Ollerer"}
}
\value{
A vector of numeric scores per observation (row).
}
\description{
Hotelling T2 scoring method for calculating multivariate anomaly scores.
}
\examples{
data <- LambsWeight[58:62,]

#Create score per sensor
T2.score<-T2(t(data))

#View scores with MultiNav function
anomaly.scores<- cbind(id=as.numeric(colnames(data)), T2.score)
MultiNav(data,type = "scores", scores = anomaly.scores)

}
\references{
ROUSSEEUW, P. J., et al. Robustbase: basic robust statistics. R package version 0.4-5, URL http://CRAN. R-project. org/package= robustbase, 2009.

Montgomery, Douglas C. Introduction to statistical quality control. John Wiley & Sons, 2007.
}
